// web/src/types/program.type.ts
export interface ProgramProps {
  program_id: string;
  name: string;
  description: string;
  max_participants: number;
  is_published: boolean;
  picture_url: string;
  start_time: string;
  end_time: string;
  open_registration: string;
  close_registration: string;
  created_at: string;
  updated_at: string;
}